Update_Chuva-main/google_functions/update_dados/inmet/requirements.txt
flask
pandas
selenium
google-cloud-bigquery
gunicorn
